Stability improvements and bug research
There seems to be a bug in OpenNI that stops the flow of frames when
using more than one sensor at a time. My current suspicion is that it
appears when the cameras are connected to the same USB bus.
I have now switched from polling for frames to using the event based
system. This can potentially cause issues with simultanious access from
different threads, so a mutex lock should be added if tearing or other
sync issues becomes apparent.