Make MM default to no classifier. Optimize.
Optimization was to pull the voxel computation out of the inner loop, to
give the compiler an easier job optimizing. This also makes it easier
for us to optimize later as well.

Defaulting to no classifier and skipping the trilinear interpolation for
calculating the background probability also speeds things up for
applications that only cares about the point clouds and/or the voxel
data, and not BG subtraction.